Refactor e2e stack setup into modules
Extract common e2e helpers into separate modules and simplify the test. Added tests/e2e/fixtureSettings.js to parse fixture settings, tests/e2e/panDomainCookie.js to create pan-domain cookies from fixture private keys, and tests/e2e/stackContainers.js to encapsulate Docker image builds and Testcontainers orchestration (startLocalStack/stopLocalStack). Updated tests/e2e/stack.spec.js to use the new helpers, removing duplicated setup/teardown and improving separation of concerns while preserving existing behaviour and error cleanup.
